+++ tests/manual/test-do.dtq.log.R
@@ -8,26 +8,26 @@ test_that("do.dtq.log exclude-include opts", {
           "dtq.log.include" = character())
   dtl(purge = TRUE)
   DT[]
-  expect_identical(dtq.log$length(), 0L, info="exclude log global env")
+  expect_identical(.DTQ$length(), 0L, info="exclude log global env")
   
   options("dtq.log.exclude" = character(),
           "dtq.log.include" = "dwtools")
   dtl(purge = TRUE)
   DT[]
-  expect_identical(dtq.log$length(), 0L, info="include log from ext package")
+  expect_identical(.DTQ$length(), 0L, info="include log from ext package")
   
   options("dtq.log.exclude" = character(),
           "dtq.log.include" = "R_GlobalEnv")
   dtl(purge = TRUE)
   DT[]
-  writeLines(dtq.log$log[[1L]]$env,"env.txt")
-  expect_identical(dtq.log$length(), 1L, info="include log global env")
+  writeLines(.DTQ$log[[1L]]$env,"env.txt")
+  expect_identical(.DTQ$length(), 1L, info="include log global env")
   
   options("dtq.log.exclude" = "R_GlobalEnv",
           "dtq.log.include" = "R_GlobalEnv")
   dtl(purge = TRUE)
   DT[]
-  expect_identical(dtq.log$length(), 0L, info="include and exclude log global env")
+  expect_identical(.DTQ$length(), 0L, info="include and exclude log global env")
   
   options("dtq.log.exclude" = character(),
           "dtq.log.include" = character())